LayoutTests/http/tests/referrer-policy/strict-origin-when-cross-origin/cross-origin-http-UpgradeMixedContent.https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/referrer-policy/strict-origin-when-cross-origin/cross-origin-http-UpgradeMixedContent.https.html requested insecure content from http://localhost:8000/referrer-policy/resources/document.html. This content was blocked and must be served over HTTPS.

Tests the behavior of strict-origin-when-cross-origin referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Timeout reached, insecure iframe not loaded
PASS successfullyParsed is true

TEST COMPLETE
